moved action bar creation into main activity
