+++ Moodlet/app/src/main/java/com/cmput3owo1/moodlet/activities/MainActivity.java
@@ -1,10 +1,15 @@
 package com.cmput3owo1.moodlet.activities;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
 
 import com.cmput3owo1.moodlet.R;
+import com.cmput3owo1.moodlet.services.UserService;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.navigation.NavController;
@@ -18,6 +23,8 @@ import androidx.navigation.ui.NavigationUI;
  */
 public class MainActivity extends AppCompatActivity {
 
+
+    private UserService userService;
     /**
      * Called when the activity is starting. Sets up the navigation.
      * @param savedInstanceState Used to restore an activity's previous state
@@ -26,6 +33,8 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        userService = new UserService();
         BottomNavigationView navView = findViewById(R.id.nav_view);
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
@@ -40,5 +49,35 @@ public class MainActivity extends AppCompatActivity {
         NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
         NavigationUI.setupWithNavController(navView, navController);
 
+    }
+
+    /** Initialize the contents of the Activity's standard options menu.
+     * @param menu The options menu in which you place your items.
+     */
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.general_fragment_menu, menu);
+        return super.onCreateOptionsMenu(menu);
+    }
+
+    /**
+     * A hook that is called whenever an item in the options menu is selected.
+     * This method can be used for adding functionality for specific selected menu items.
+     * @param item The menu item that was selected
+     * @return
+     */
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+
+        switch(item.getItemId()){
+            case R.id.logout:
+                userService.logoutUser();
+                Intent intent = new Intent(this, LoginActivity.class);
+                intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
+                startActivity(intent);
+                finish();
+        }
+
+        return super.onOptionsItemSelected(item);
     }
 }
