+++ Moodlet/app/src/main/java/com/cmput3owo1/moodlet/fragments/FeedFragment.java
@@ -1,11 +1,7 @@
 package com.cmput3owo1.moodlet.fragments;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ListView;
@@ -15,10 +11,10 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import com.cmput3owo1.moodlet.R;
-import com.cmput3owo1.moodlet.activities.LoginActivity;
 import com.cmput3owo1.moodlet.adapters.FeedListAdapter;
 import com.cmput3owo1.moodlet.models.MoodEventAssociation;
 import com.cmput3owo1.moodlet.services.IMoodEventServiceProvider;
+import com.cmput3owo1.moodlet.services.IUserServiceProvider;
 import com.cmput3owo1.moodlet.services.MoodEventService;
 import com.cmput3owo1.moodlet.services.UserService;
 
@@ -34,7 +30,7 @@ public class FeedFragment extends Fragment implements IMoodEventServiceProvider.
     private FeedListAdapter feedAdapter;
     private ArrayList<MoodEventAssociation> feedDataList;
     private IMoodEventServiceProvider service;
-    private UserService userService;
+    private IUserServiceProvider userService;
 
     /**
      * Called when the fragment is starting.
@@ -70,32 +66,6 @@ public class FeedFragment extends Fragment implements IMoodEventServiceProvider.
         userService = new UserService();
 
         return rootView;
-    }
-
-    /**
-     * Initialize the contents of the Activity's standard options menu.
-     * @param menu The options menu in which you place your items.
-     * @param inflater The MenuInflater object that can be used to inflate any itmes in the menu
-     */
-    @Override
-    public void onCreateOptionsMenu(@NonNull Menu menu, @NonNull MenuInflater inflater) {
-        inflater.inflate(R.menu.general_fragment_menu, menu);
-        super.onCreateOptionsMenu(menu, inflater);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-
-        switch(item.getItemId()){
-            case R.id.logout:
-                userService.logoutUser();
-                Intent intent = new Intent(getActivity(), LoginActivity.class);
-                intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
-                startActivity(intent);
-                getActivity().finish();
-        }
-
-        return super.onOptionsItemSelected(item);
     }
 
     /**
